requirements_aws.txt

# requirements_aws.txt
# ---------- Base tools ----------
pip>=24.0
setuptools>=70.0
wheel>=0.44.0

# ---------- Core Deep Learning stack ----------
# CUDA-enabled torch/vision (CUDA 12.1 wheels)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121
# ---------- Utility and training helpers ----------
torchinfo>=1.8.0
torch-lr-finder>=0.2.1,<0.3.0
fire>=0.6.0

# ---------- Image / Augmentation ----------
timm==1.0.9
albumentations==1.4.14
albucore==0.0.13
opencv-python==4.10.0.84
Pillow>=10.3.0

# ---------- Utilities ----------
tqdm==4.66.4
wandb==0.17.5
yacs==0.1.8
matplotlib>=3.9.2
numpy>=1.26.4

# ---------- Optional (safe to keep) ----------
scikit-learn==1.5.2
pandas>=2.2.2

# optional pipeline speedup:
nvidia-dali-cuda120
